Split up XML classes one per file
