+++ src/JuliusSweetland.OptiKey/JuliusSweetland.OptiKey.csproj
@@ -243,6 +243,10 @@
     <Compile Include="Models\DynamicKeyboardFolder.cs" />
     <Compile Include="Models\ILastMouseActionStateManager.cs" />
     <Compile Include="Models\KeyboardOutputServiceState.cs" />
+    <Compile Include="Models\XmlKeyboardModels\XmlActionKey.cs" />
+    <Compile Include="Models\XmlKeyboardModels\XmlChangeKeyboardKey.cs" />
+    <Compile Include="Models\XmlKeyboardModels\XmlGrid.cs" />
+    <Compile Include="Models\XmlKeyboardModels\XmlKey.cs" />
     <Compile Include="Models\XmlKeyboardModels\XmlKeyboard.cs" />
     <Compile Include="Models\KeyStateServiceState.cs" />
     <Compile Include="Models\KeyPressKeyValue.cs" />
@@ -257,6 +261,9 @@
     <Compile Include="Models\NotificationWithServicesAndState.cs" />
     <Compile Include="Models\NotificationEventArgs.cs" />
     <Compile Include="Models\SerializableDictionaryOfTimeSpanByKeyValues.cs" />
+    <Compile Include="Models\XmlKeyboardModels\XmlKeys.cs" />
+    <Compile Include="Models\XmlKeyboardModels\XmlTextKey.cs" />
+    <Compile Include="Models\XmlKeyboardModels\XmlUtils.cs" />
     <Compile Include="Native\Common\Enums\AppBarEdge.cs" />
     <Compile Include="Native\Common\Enums\AppBarMessages.cs" />
     <Compile Include="Native\Common\Enums\AppBarNotify.cs" />
