+++ src/JuliusSweetland.OptiKey/Models/XmlKeyboardModels/XmlGrid.cs
@@ -1,0 +1,11 @@
+namespace JuliusSweetland.OptiKey.Models
+{
+    public class XmlGrid
+    {
+        public int Rows
+        { get; set; }
+
+        public int Cols
+        { get; set; }
+    }
+}
