+++ src/JuliusSweetland.OptiKey/Models/XmlKeyboardModels/XmlTextKey.cs
@@ -1,0 +1,8 @@
+namespace JuliusSweetland.OptiKey.Models
+{
+    public class XmlTextKey : XmlKey
+    {
+        public string Text
+        { get; set; }
+    }
+}
